Followed runtime-editor split
