+++ test/editor/pageobjects/nodes/core/logic/15-change_page.js
@@ -54,8 +54,8 @@ changeNode.prototype.ruleSet = function(p, pt, to, tot, index) {
     setT("set", index);
     if (pt) {
         browser.clickWithWait('//*[@id="node-input-rule-container"]/li[' + index + ']/div/div[1]/div/button[1]');
-        var num = 5 * index + 6;
-        var ptXPath = '/html/body/div[' + num + ']/a[' + ptType[pt] + ']';
+        var num = 5 * (index - 1) + 1;
+        var ptXPath = '//div[@class="red-ui-typedInput-options"][' + num + ']/a[' + ptType[pt] + ']';
         browser.clickWithWait(ptXPath);
     }
     if (p) {
@@ -63,8 +63,8 @@ changeNode.prototype.ruleSet = function(p, pt, to, tot, index) {
     }
     if (tot) {
         browser.clickWithWait('//*[@id="node-input-rule-container"]/li[' + index + ']/div/div[2]/div[2]/button[1]');
-        var num = 5 * index + 7;
-        var totXPath = '/html/body/div[' + num + ']/a[' + totType[tot] + ']';
+        var num = 5 * (index - 1) + 2;
+        var totXPath = '//div[@class="red-ui-typedInput-options"][' + num + ']/a[' + totType[tot] + ']';
         browser.clickWithWait(totXPath);
     }
     if (to) {
